refactor: improve readability of the Makefile
This commit tries to improve the readability of the Makefile by combining
the two Makefiles, one from `deployments/k3d` and another from the root
directory into one. It adds more comments and simplified / deleted
commands. Hopefully it will make the usage much easier.

It also adds guidance on how to build / run the Spin shim locally on the
host, in addition to that of on Kubernetes.

Signed-off-by: jiaxiao zhou <jiazho@microsoft.com>